Use checks API in addition to statuses API to determine PR CI status
The script currently fails to see checks that do not use the statuses API, e.g., TravisCI.com, and is/was subsequently merging PRs failing CI. This should fix that.

Note: I was not able to test this against lots of repos, so I will work with the FR next week to make sure this works properly with repos that DO NOT use checks. cc: @justinlittman